refactor: Aggregate all filters into filter.go
