--- packages/tangle/filter_message.go
@@ -1,16 +1,0 @@
-package tangle
-
-import (
-	"github.com/iotaledger/hive.go/autopeering/peer"
-)
-
-// MessageFilter filters based on messages and peers.
-type MessageFilter interface {
-	// Filter filters up on the given message and peer and calls the acceptance callback
-	// if the input passes or the rejection callback if the input is rejected.
-	Filter(msg *Message, peer *peer.Peer)
-	// OnAccept registers the given callback as the acceptance function of the filter.
-	OnAccept(callback func(msg *Message, peer *peer.Peer))
-	// OnAccept registers the given callback as the rejection function of the filter.
-	OnReject(callback func(msg *Message, err error, peer *peer.Peer))
-}
